+++ web/src/savedSearches/SavedSearchListPage.tsx
@@ -157,11 +157,7 @@ export class SavedSearchListPage extends React.Component<Props, State> {
                         !isErrorLike(this.state.savedSearchesOrError) &&
                         this.state.savedSearchesOrError.length > 0 &&
                         this.state.savedSearchesOrError
-                            .filter(
-                                search =>
-                                    search.orgID === this.props.namespace.id ||
-                                    search.userID === this.props.namespace.id
-                            )
+                            .filter(search => this.props.namespace.id === search.namespace.id)
                             .map(search => (
                                 <SavedSearchNode
                                     key={search.id}